Make Utils package protected and rename them
