+++ impl/maven-core/src/main/java/org/apache/maven/internal/impl/DefaultProject.java
@@ -47,7 +47,7 @@ import org.apache.maven.project.MavenProject;
 import org.apache.maven.project.artifact.ProjectArtifact;
 import org.eclipse.aether.util.artifact.ArtifactIdUtils;
 
-import static org.apache.maven.impl.Utils.nonNull;
+import static org.apache.maven.internal.impl.CoreUtils.nonNull;
 
 public class DefaultProject implements Project {
