+++ impl/maven-impl/src/main/java/org/apache/maven/impl/DefaultArtifactInstaller.java
@@ -29,7 +29,7 @@ import org.eclipse.aether.RepositorySystem;
 import org.eclipse.aether.installation.InstallRequest;
 import org.eclipse.aether.installation.InstallationException;
 
-import static org.apache.maven.impl.Utils.nonNull;
+import static org.apache.maven.impl.ImplUtils.nonNull;
 
 @Named
 @Singleton
