+++ impl/maven-impl/src/main/java/org/apache/maven/impl/DefaultChecksumAlgorithmService.java
@@ -41,7 +41,7 @@ import org.apache.maven.api.services.ChecksumAlgorithmServiceException;
 import org.eclipse.aether.spi.connector.checksum.ChecksumAlgorithmFactory;
 import org.eclipse.aether.spi.connector.checksum.ChecksumAlgorithmFactorySelector;
 
-import static org.apache.maven.impl.Utils.nonNull;
+import static org.apache.maven.impl.ImplUtils.nonNull;
 
 @Named
 @Singleton
