+++ impl/maven-impl/src/main/java/org/apache/maven/impl/DefaultRepositoryFactory.java
@@ -32,7 +32,7 @@ import org.apache.maven.api.services.RepositoryFactory;
 import org.eclipse.aether.impl.RemoteRepositoryManager;
 import org.eclipse.aether.repository.RepositoryPolicy;
 
-import static org.apache.maven.impl.Utils.nonNull;
+import static org.apache.maven.impl.ImplUtils.nonNull;
 
 @Named
 @Singleton
